requirements.txt
fastapi~=0.116.1
uvicorn[standard]
streamlit~=1.49.1
langchain~=0.3.27
langchain-groq
groq~=0.31.1
chromadb
PyPDF2~=3.0.1
python-docx~=1.2.0
sendgrid~=6.12.4
python-multipart
python-dotenv~=1.1.1
requests~=2.32.5
python-json-logger
prometheus-client
pytest

pydantic~=2.11.9
langchain-community~=0.3.29
transformers~=4.56.1
certifi~=2025.8.3